# cryptostockz-back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "CryptostockZ"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:10010"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
...
  /product:
    post:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Add a new product to the store"
      description: ""
      operationId: "createProduct"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        "405":
          description: "Invalid input"
    put:
      x-swagger-router-controller: "product.controller"
      tags:
      - "product"
      summary: "Update an existing product"
      description: ""
      operationId: "updateProduct"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Product not found"
        "405":
          description: "Validation exception"
...
  /account/{username}:
    get:
      x-swagger-router-controller: "account.controller"
      tags:
      - "account"
      summary: "Get user by user name"
      description: ""
      operationId: "getUserByName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
    put:
      x-swagger-router-controller: "account.controller"
      tags:
      - "account"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "400":
          description: "Invalid user supplied"
        "404":
          description: "User not found"
    delete:
      x-swagger-router-controller: "account.controller"
      tags:
      - "account"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
  /account/products/all:
    get:
      x-swagger-router-controller: "account.controller"
      tags:
      - "account"
      summary: "Finds user owned Products"
      description: ""
      operationId: "getUserProducts"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
        "400":
          description: "Invalid status value"
...
  /manufacturers:
    get:
      x-swagger-router-controller: "account.controller"
      tags:
      - "account"
      summary: "Finds all manufacturers"
      description: ""
      operationId: "getManufacturers"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
        "400":
          description: "Invalid status value"

  #Operaciones sobre productos base
...
definitions:
  User:
    type: "object"
    required:
    - "username"
    - "email"
    - "password"
    properties:
      username:
        type: "string"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      role:
        type: "integer"
        enum:
          - 1
          - 2
          - 3
          - 4
      purchases:
        type: "number"
      sales:
        type: "number"
      level:
        type: "integer"
      permisions:
        type: "integer"
        enum:
          - 1
          - 2
      metamaskAccount:
        type: "string"
    xml:
      name: "User"
  Product:
    type: "object"
    properties:
      address:
        type: "string"
      base_product_id:
        type: "integer"
      unique_identificator:
        type: "string"
      manufact_id:
        type: "string"
      owner_address:
        type: "string"
      level:
        type: "integer"
    xml:
      name: "Product"
...